+++ src/components/index.ts
@@ -1,4 +1,1 @@
 export { Avatar } from "./Avatar";
-export { useToast } from "./Layout/Toast";
-
-export * as Icons from "./Icons";
