+++ src/hooks/use-online.ts
@@ -1,6 +1,6 @@
 import { useCallback, useState } from "react";
 
-import { useToast } from "components";
+import { useToast } from "components/Toast";
 import { NETWORK_STATE } from "utils/constants";
 import { useIsomorphicLayoutEffect } from "hooks";
