+++ src/pages/staff-teacher/subjects/create.tsx
@@ -5,7 +5,7 @@ import useSWR from "swr";
 import Head from "next/head";
 import dynamic from "next/dynamic";
 
-import { useToast } from "components";
+import { useToast } from "components/Toast";
 import { fetchAPIEndpoint } from "api/client";
 import { LoadingIcon } from "components/Icons";
 import { useIsomorphicLayoutEffect } from "hooks";
